Update Boston example
