+++ boston/tensorflow/keras/README.md
@@ -121,6 +121,10 @@ packages/dependencies and set a few options.
 from setuptools import find_packages
 from setuptools import setup
 
+with open('requirements.txt') as f:
+  requirements = [l.strip('\n') for l in f if
+                  l.strip('\n') and not l.startswith('#')]
+                  
 setup(
   name='boston',
   version='0.1',
