+++ boston/tensorflow/keras/setup.py
@@ -2,6 +2,9 @@
 from setuptools import find_packages
 from setuptools import setup
 
+with open('requirements.txt') as f:
+  requirements = [l.strip('\n') for l in f if
+                  l.strip('\n') and not l.startswith('#')]
 
 setup(
   name='boston',
